% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Budget.R
\name{Budget}
\alias{Budget}
\title{Budget}
\usage{
Budget(i, e, s)
}
\arguments{
\item{i}{The income amount}

\item{e}{The expenses amount}

\item{s}{The savings amount}
}
\value{
amount to spend on fun
}
\description{
Budget
}
\examples{
Budget(6000,500,2000)
}
